Improve yeelight component (#22347)
